fix: tdp-cluster var display
Fixes 118
