+++ src/components/Services/ComponentsNav/type.ts
@@ -1,4 +1,4 @@
-export type Tab = {
+export type TTab = {
   id: string
   href: string
 }
