+++ src/components/Services/index.tsx
@@ -1,1 +1,2 @@
+export * from './ComponentsNav'
 export * from './VariablesDisplay'
